Move block providers to helper functions
